gnbutk yhri

+++ react_project/src/component/comment.jsx
@@ -21,9 +21,17 @@ const Comment = ({ postId }) => {
 
     const deleteComment = async (id) => {
         try {
-            const response = await axios.delete(`http://localhost:5300/comment/${id}`);
-            console.log(response);
-            removeCommentById(id);
+            const user = localStorage.getItem('currentUser');
+            let currentUser = JSON.parse(user)
+
+            if (getUserIdFromCommentId(id) === currentUser.id_user) {
+                const response = await axios.delete(`http://localhost:5300/comment/${id}`);
+                
+                console.log(response);
+                removeCommentById(id);
+            } else {
+                alert("This comment does not belong to you!!!");
+            }
         } catch (error) {
             console.error('Error occurred during authentication:', error);
         }
@@ -33,12 +41,21 @@ const Comment = ({ postId }) => {
         setComment(prevComments => prevComments.filter(comment => comment.id_comment !== id));
     };
 
-    const addComment = async ({ postId}) => {
+    const addComment = async ({ postId }) => {
         try {
             const response = await axios.post(`http://localhost:5300/comment`, { name: name, content: content, mail: mail, postId: postId }); // הוספת post_id לבקשה
             // const response2 = await axios.get(`http://localhost:5300/comment`);
             console.log(response);
             setComment([...comment, response.data]);
+        } catch (error) {
+            console.error('Error occurred during authentication:', error);
+        }
+    };
+    const getUserIdFromCommentId = async ({ commentId }) => {
+        try {
+            const response = await axios.get(`http://localhost:5300/comment/userIdFromComment`, { comment_id: commentId });
+            console.log(response.data);
+            return response.data.userID;
         } catch (error) {
             console.error('Error occurred during authentication:', error);
         }
